test: explicitly cast to uint32_t in test
Explicitly casting to `uint32_t` prevents a template resolution
ambiguity on 32-bit platforms.

Fixes: https://github.com/nodejs/node-addon-api/issues/337

PR-URL: https://github.com/nodejs/node-addon-api/pull/341
Fixes: https://github.com/nodejs/node-addon-api/issues/337

+++ test/object/object.cc
@@ -126,7 +126,7 @@ Value CreateObjectUsingMagic(const CallbackInfo& info) {
   obj[std::string("s_true")] = true;
   obj[std::string("s_false")] = false;
   obj["0"] = 0;
-  obj[42] = 120;
+  obj[(uint32_t)42] = 120;
   obj["0.0f"] = 0.0f;
   obj["0.0"] = 0.0;
   obj["-1"] = -1;
